Hk issue 37 (#38)
Fixes #37. That one took me three days of looking at and thinking about and 1 second to fixed once I had it reproduced lol. Due to less than ideal coverage of the "inject subject into specs file", I accidentally moved the node.string = "${idx}" into the load files loop, meaning the require string was not replaced if the weren't any files to load. This in turn lead to a compile error when running crytic on my blog, which is somehow masked because compile errors appear to be killed mutants. I think I will output any compile errors even for killed mutants, because only like this was I able to spot the bug.