Ooof. Be more lax about the expected output.
We have a default preamble which enables
fail fast mode in specs. Between crystal 0.30.1
and 0.31.0 this changed the output. The new crystal
doesnt print a summary anymore, which means we will
not find "Finished" in it. Instead only one "F" is printed and
that's it (if there is exactly one test which fails).

See also https://github.com/crystal-lang/crystal/issues/8359